requirements.txt
fastapi[standard]
RPi.GPIO
